phi/gfx: pass RenderFx to Renderable and CopySprite trait's methods
When rendering sprites we need a way to tell the renderer to apply some
effects like flipping horizontally.

Introduce therefore the RenderFx enum with the following elements:

  * FlipX	flip on the y axis (horizontally)
  * None	don't do anything

This enumeration should be populated later with more effects.

The enum is then passed to render() of the Renderable trait and
copy_sprite() of the CopySprite trait.

The classes implementing those traits are made aware on how to apply
those effects.

The calling code is then changed accordingly.