phi/mod.rs: Fix the lifetime issues related to font handling
The issue stems from the fact that the Phi struct needs a Sdl2TtfContext
to load new fonts.

Therefore the ttf context needs to be passed to Phi as a non-mutable
reference in its new() method, so add it as a third argument of
Phi::new().

Since Phi is a struct, the lifetime of the Sdl2TtfContext needs to be
explicit and should match the lifetime of the Fonts in its HashMap.

The signature of Phi::new() is updated accordingly.